# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ante",
    "description": "AI-Native Trading Engine API",
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/config": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "List Configs",
        "description": "동적 설정 전체 조회. 인증된 master/human 또는 ``config:read`` scope 를\n보유한 agent 만 호출 가능 (#1407).",
        "operationId": "list_configs_api_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigListResponse"
                }
              }
            }
          },
          "503": {
            "description": "Config service not available",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/config/{key}": {
      "put": {
        "tags": [
          "config"
        ],
        "summary": "Update Config",
        "description": "동적 설정 값 변경. 인증된 master/human 또는 config:write scope 보유\nagent 만 호출 가능 (#1373).\n\n``update_bot`` (#1352) / ``create_bot`` (#1371) / ``allocate`` (#1372)\n와 동일한 raw body 파싱 패턴을 적용해 인증 가드가 body validation 보다\n우선 실행되도록 한다. FastAPI 가 ``body: ConfigUpdateRequest`` 를 먼저\n검증하면 unauth + bad-body 시 401 이 아닌 422 가 먼저 반환되어 contract\n가 깨진다.\n\n핸들러 단계 순서:\n\n1. 인증 가드 (``Depends(require_config_write)``) — caller 빈 → 401,\n   권한 없음 → 403, 비활성 멤버 → 403.\n2. raw bytes 읽기 + JSON 파싱 — 실패 시 422.\n3. ``ConfigUpdateRequest.model_validate`` — ValidationError → 422.\n4. config 키 존재 확인 → 404.\n5. ``config_service.set`` 호출 + audit 기록 (changed_by = caller_id).\n   서비스 경계에서 ``validate_value`` 가 키별 invariant 를 검증하므로\n   ``ValueError`` 가 올라오면 422 로 변환한다 (#1379).",
        "operationId": "update_config_api_config__key__put",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigUpdateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required (missing or invalid Authorization header AND missing or invalid ante_session cookie). 대시보드 사용자는 로그인 후 ante_session 쿠키만 가지고 호출하며, 에이전트 클라이언트는 Bearer 토큰만 가지고 호출한다. 둘 중 하나라도 유효하면 통과한다.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied (master, human 멤버 또는 config:write scope 보유 agent 만 허용)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Config key not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Body validation 실패 (JSON 파싱 실패, 빈 body, value 누락, type mismatch) 또는 키별 invariant 위반 (예: ``system.log_level`` 가 ``_VALID_LOG_LEVELS`` 멤버가 아닌 경우 — 대소문자 구분, #1379). 단, 인증이 실패하면 body validation은 실행되지 않고 401이 우선 반환된다(#1373).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Config service not available",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigUpdateRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/reports": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Submit Report",
        "description": "리포트 제출. 인증된 master/human 또는 ``report:write`` scope 를 보유한\nagent 만 호출 가능 (#1374).\n\n``update_config`` (#1373) 와 동일한 raw body 파싱 패턴을 적용해 인증 가드가\nbody validation 보다 우선 실행되도록 한다. FastAPI 가\n``body: ReportSubmitRequest`` 를 먼저 검증하면 unauth + bad-body 시 401 이\n아닌 422 가 먼저 반환되어 contract 가 깨진다.\n\n핸들러 단계 순서:\n\n1. 인증 가드 (``Depends(require_report_write)``) — caller 빈 → 401,\n   권한 없음 → 403, 비활성 멤버 → 403.\n2. raw bytes 읽기 + JSON 파싱 — 실패 시 422.\n3. ``ReportSubmitRequest.model_validate`` — ValidationError → 422.\n4. ``report_store.submit`` 호출 + audit 기록 (submitted_by = caller_id).",
        "operationId": "submit_report_api_reports_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSubmitResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required (missing or invalid Authorization header AND missing or invalid ante_session cookie). 대시보드 사용자는 로그인 후 ante_session 쿠키만 가지고 호출하며, 에이전트 클라이언트는 Bearer 토큰만 가지고 호출한다. 둘 중 하나라도 유효하면 통과한다.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied (master, human 멤버 또는 report:write scope 보유 agent 만 허용). spec require_scope predicate 정합.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Body validation 실패 (JSON 파싱 실패, 빈 body, 필수 필드 누락, type mismatch, NaN/Inf, extra key). 단, 인증이 실패하면 body validation 은 실행되지 않고 401 이 우선 반환된다(#1374).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Report store not available",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportSubmitRequest"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List Reports",
        "description": "리포트 목록 조회 (cursor 기반 페이지네이션). 인증된 master/human 또는\n``report:read`` scope 를 보유한 agent 만 호출 가능 (#1407).",
        "operationId": "list_reports_api_reports_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReportStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportListResponse"
                }
              }
            }
          },
          "503": {
            "description": "Report store not available",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/{report_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Report View",
        "description": "리포트 단건 조회. 인증된 master/human 또는 ``report:read`` scope 를\n보유한 agent 만 호출 가능 (#1407).",
        "operationId": "report_view_api_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Report not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Report store not available",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigItem": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "title": "Value"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": ""
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "key"
        ],
        "title": "ConfigItem",
        "description": "동적 설정 아이템."
      },
      "ConfigListResponse": {
        "properties": {
          "configs": {
            "items": {
              "$ref": "#/components/schemas/ConfigItem"
            },
            "type": "array",
            "title": "Configs"
          }
        },
        "type": "object",
        "required": [
          "configs"
        ],
        "title": "ConfigListResponse",
        "description": "동적 설정 목록 응답."
      },
      "ConfigUpdateResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "old_value": {
            "title": "Old Value"
          },
          "new_value": {
            "title": "New Value"
          }
        },
        "type": "object",
        "required": [
          "key"
        ],
        "title": "ConfigUpdateResponse",
        "description": "설정 변경 응답."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportListItem": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "title": "Submitted At"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "strategy",
          "status",
          "submitted_at"
        ],
        "title": "ReportListItem",
        "description": "리포트 목록 아이템."
      },
      "ReportListResponse": {
        "properties": {
          "reports": {
            "items": {
              "$ref": "#/components/schemas/ReportListItem"
            },
            "type": "array",
            "title": "Reports"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "reports"
        ],
        "title": "ReportListResponse",
        "description": "리포트 목록 응답."
      },
...
      "ReportStatus": {
        "type": "string",
        "enum": [
          "draft",
          "submitted",
          "reviewed",
          "adopted",
          "rejected",
          "archived"
        ],
        "title": "ReportStatus",
        "description": "리포트 상태."
      },
      "ReportSubmitResponse": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "strategy",
          "status"
        ],
        "title": "ReportSubmitResponse",
        "description": "리포트 제출 응답."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ErrorResponse": {
        "description": "RFC 7807 Problem Details 에러 응답.",
        "properties": {
          "type": {
            "default": "/errors/internal",
            "title": "Type",
            "type": "string"
          },
          "title": {
            "default": "Internal Server Error",
            "title": "Title",
            "type": "string"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "status": {
            "default": 500,
            "title": "Status",
            "type": "integer"
          },
          "instance": {
            "default": "",
            "title": "Instance",
            "type": "string"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ConfigUpdateRequest": {
        "type": "object",
        "title": "ConfigUpdateRequest",
        "description": "PUT /api/config/{key} 입력 contract. 인증된 master/human 또는 config:write scope 를 보유한 agent 만 사용할 수 있다(#1373). Bearer 토큰 또는 유효한 ante_session 쿠키 중 하나라도 있어야 하며, 둘 다 없거나 둘 다 invalid면 body validation 전에 401로 차단된다.",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "description": "변경할 설정 값. 타입은 키마다 다르다 — DynamicConfigService 가 키별 schema 검증을 담당한다."
          },
          "category": {
            "type": "string",
            "default": "",
            "description": "설정 카테고리. 비우면 key 의 dot-prefix(예: ``risk.max_mdd`` → ``risk``)에서 자동 추출된다."
          }
        }
      },
      "ReportSubmitRequest": {
        "additionalProperties": false,
        "description": "리포트 제출 요청.\n\nSSOT: ``docs/specs/report-store/report-store.md`` 리포트 제출 스키마.\n``win_rate``는 percent 단위 (0.0~100.0, ``\"win_rate\": 58.0`` 예시)다.\n\ninvariants (#1353):\n- ``total_trades >= 0`` (거래 수는 음수 불가).\n- ``win_rate``는 ``[0.0, 100.0]`` percent 단위 또는 ``None``.\n- ``total_return_pct``/``sharpe_ratio``/``max_drawdown_pct``/``win_rate``는\n  finite. NaN/Inf는 후속 분석/표시에서 버그 또는 잘못된 결정으로 이어지므로\n  submit 단에서 422로 거부한다.\n- 미지정 키는 ``extra='forbid'``로 거부 — agent의 오타/잘못된 키가 store에\n  유실되는 대신 빠르게 fail하도록 한다.\n\nNote: 단위 통일(percent ↔ ratio)은 detail_json·backtest_runs·CLI submit 등\n시스템 전반의 영향이 커서 본 PR에서는 input validation에만 집중하고,\nratio 통일은 별도 follow-up 이슈에서 다룬다.",
        "properties": {
          "strategy_name": {
            "title": "Strategy Name",
            "type": "string"
          },
          "strategy_version": {
            "title": "Strategy Version",
            "type": "string"
          },
          "strategy_path": {
            "title": "Strategy Path",
            "type": "string"
          },
          "backtest_period": {
            "default": "",
            "title": "Backtest Period",
            "type": "string"
          },
          "total_return_pct": {
            "default": 0.0,
            "title": "Total Return Pct",
            "type": "number"
          },
          "total_trades": {
            "default": 0,
            "minimum": 0,
            "title": "Total Trades",
            "type": "integer"
          },
          "sharpe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpe Ratio"
          },
          "max_drawdown_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Drawdown Pct"
          },
          "win_rate": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Win Rate"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "rationale": {
            "default": "",
            "title": "Rationale",
            "type": "string"
          },
          "risks": {
            "default": "",
            "title": "Risks",
            "type": "string"
          },
          "recommendations": {
            "default": "",
            "title": "Recommendations",
            "type": "string"
          },
          "detail_json": {
            "default": "{}",
            "title": "Detail Json",
            "type": "string"
          },
          "sections": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sections"
          }
        },
        "required": [
          "strategy_name",
          "strategy_version",
          "strategy_path"
        ],
        "title": "ReportSubmitRequest",
        "type": "object"
      },
...
    }
  }
}